requirements.txt
asgiref==3.2.10
Django==3.0.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytest
pytest-django
PyJWT==1.7.1
pytz==2020.1
python-dotenv
requests
sqlparse==0.3.1
django-extensions